Added a boolean parameter 'failFastOnNull' to the Build method on both the ValidationBuilder and TenantValidationBuider set to false by default. Its purpose when set to true is to return a single InvalidEntry with a failure message stating that the entity used is null, stopping all validations otherwise multiple failure messages accrue.
Added an extension method HasNullEntityFailure on Validated<T> that checks if there are failures and returns true if there is a null entity failure message.

Added support for conditional validation on the ValidationBuilder by adding the Methods DoWhen and EndWhen. All validators after the DoWhen are included in the conditional until an EndWhen closes the group.

Added the extension method on EntityValidator<TEntity> to provide the conditional predicate needed for DoWhen.