Added code so nested conditional scopes with DoWhen/EndWhen is now supported.
The ValidationBuilder.Build method now throws InvalidOperationException if DoWhen calls are not matched with a corresponding EndWhen.

Bumped version to 1.1.1 ready for patch release. Classing as a patch as no exception was raised in V1.1 if there were unbalanced scopes which may not have yielded predicable results as everything would have been in scope without the EndWhen.